Project-Stages/Project-Stage-1-Basic-CI-CD-Deploy/Stage-1-Architecture/requirements.txt
# Requirements for Stage 1 Architecture Diagrams
# Health Care Management System - Diagram as Code

# Core diagram generation library
diagrams>=0.23.3

# Image processing library (required by diagrams)
Pillow>=10.0.0

# Optional: For enhanced diagram features
graphviz>=0.20.1

# Optional: For diagram customization
matplotlib>=3.7.0
